apple/apple2gs: Restrict DOCRAM pointer to 16 bits like hardware, Great Western Shootout relies on it wrapping. [R. Belmont, Peter Ferrie]

+++ src/mame/apple/apple2gs.cpp
@@ -447,7 +447,7 @@ private:
 
 	// Sound GLU variables
 	u8 m_sndglu_ctrl = 0;
-	int m_sndglu_addr = 0;
+	u16 m_sndglu_addr = 0;
 	int m_sndglu_dummy_read = 0;
 
 	// Key GLU variables
